consolidate llama, falcon, mixtral kv caches
